==> book/2_DataRetrieval.ipynb <==
<a id="s2"></a>Data Retrieval

This section describes the process and code that we used to retrieve our raw data (from Twitter, Facebook, Bitly, and URL paths). Code for these tasks was (typically) written to be executed manually in *.py* files. Simplified versions are presented here for documentation, may not be runnable as-is without API configuration. Here, we strip out many rote tasks—loading libraries, printing status updates, etc.—in the interest of focusing on important functionality. If the reader is interested in full source code, it may be found in (and executed with) the noted files.

The following code loads all libraries utilized in this section.

from bs4 import BeautifulSoup
import datetime as dt
import dateutil.parser
from hashlib import md5
import json
import numpy as np
import os
import re
import requests
import shutil
import sys
import time
from twython import Twython
import urllib

Tweets

Recent Tweets

We use the [Twitter OAuth2 REST APIs](https://dev.twitter.com/rest/public) and [Twython package](https://twython.readthedocs.org/en/latest/) to read recent tweets from all of our target news organizations. Twitter's API imposes a 3,200 tweet limit on historical retrieval of individual accounts’ tweets, so this process simply pulled tweets in sets of 200 (an API-imposed limit) until no more remained. See below for information about retrieving older tweets. Twitter's API documentation may provides a [helpful example of the JSON data](https://dev.twitter.com/rest/reference/get/statuses/user_timeline) that that process retrieves (and store) from the API.

See the full version of this script in: [/src/twitter/downloadRawTweets.py](https://github.com/micahlanier/social-news/blob/master/src/twitter/downloadRawTweets.py).

"""

This script will pull as many recent tweets as possible from given organizations and dump them to a timestamped file.
If we pull tweets over time, this will allow us to arrange pulls in order and consolidate them into a master file.

We have encountered rate limit failures around 17 consecutive organizations.
Consider dividing the organization list in two and fetching both halves over time.
If it fails, wait 15 minutes and try again. In the future, we can think about cycling through API keys.

Relevant Twitter API documentation:
    https://dev.twitter.com/rest/reference/get/statuses/user_timeline

"""

##### Configuration

# List of organizations to fetch. Either a list or "all".
orgs = 'all'

##### Main Setup

# API setup.
apiConfigPath = '../../../conf/twitter.json'

# Data directory.
rawTweetsDirectory = '../../../data/twitter/raw/'
# Get timestamp of current run.
timestampFilename = dt.datetime.now().strftime('%Y-%m-%dT%H-%M-%S') + '.json'

# Twitter Setup
# Read API settings.
twitterSettings = json.load(open(apiConfigPath))
# Set up twitter object.
twitter = Twython(twitterSettings['apiKey'], access_token=twitterSettings['accessToken'])
# Default number of tweets to grab per request.
tweetsPerRequest = 200

# Organization setup.
allOrgs = json.load(open('../../conf/organizations.json'))
if (type(orgs) == type('str')):
    orgs = allOrgs.keys()
relevantOrgs = dict((k, allOrgs[k]) for k in orgs);

##### Get Tweets

# Iterate over organizations.
for org, orgData in relevantOrgs.iteritems():
    # Get SN.
    sn = orgData['twitter']

    # Set up container for tweets.
    tweets = []
    # Keep track of min tweet ID; min/max dates.
    minTweetId = None
    minTweetDate = None
    maxTweetDate = None
    # Set up folder if it doesn't exist.
    snDirectory = rawTweetsDirectory + org + '/'
    if not os.path.exists(snDirectory):
        os.makedirs(snDirectory)

    # Start looping and grabbing tweets as we go. Loop will terminate at a break.
    # We just set an arbitrarily high limit in case something goes haywire.
    while len(tweets) < 4000:
        # Get tweets.
        if (minTweetId is None):
            currentTweets = twitter.get_user_timeline(screen_name=sn, count=tweetsPerRequest)
        else:
            currentTweets = twitter.get_user_timeline(screen_name=sn, max_id=minTweetId, count=tweetsPerRequest)
        # If nothing there (we've reached the end), break the loop
        if (not len(currentTweets)):
            break
        # Iterate over tweets and find min ID, min/max date.
        for t in currentTweets:
            minTweetId = min([tId for tId in [minTweetId, t['id']-1] if tId is not None])
            currentTweetDate = dateutil.parser.parse(t['created_at'])
            minTweetDate = min([tDate for tDate in [minTweetDate, currentTweetDate] if tDate is not None])
            maxTweetDate = max([tDate for tDate in [maxTweetDate, currentTweetDate] if tDate is not None])
        # Append to list.
        tweets += currentTweets

    # Dump to file.
    filename = rawTweetsDirectory + org + '/' + timestampFilename
    json.dump(tweets, open(filename,'w'))

Older Tweets

Twitter's API imposes a 3,200 tweet limit on historical retrieval of individual accounts’ tweets. This extends quite far back for some users, but for more prolific tweeters (e.g., the *Huffington Post* and *New York Times*) this restricts pulls to only the last couple of weeks. In order to get a dataset that extends further back in time for the more prolific users, we devised a method for scraping Twitter's [search page](https://twitter.com/search-home) and using the API to get all information about each tweet that we found. See the full version in: [/src/twitter/scrapeOlderTweets.py](https://github.com/micahlanier/social-news/blob/master/src/twitter/scrapeOlderTweets.py).

"""

This script will scrape tweets beyond the 3200-tweet limit by scraping the Twitter web search interface.
Expect this process to be slooow.

"""

##### Configuration

# We'll use these libraries immediately.
import dateutil.parser

# Enter a single organization to scrape. This should be changed between runs.
org = 'nytimes'

# Date to go back to.
until = dateutil.parser.parse('2014/10/01')

# Number of tweets to consolidate. Output will be in files with this many tweets.
tweetsPerFile = 100

##### Main Setup

# Ensure "until" setting is comparable to minDate.
# until = pytz.utc.localize(until)
until = until.replace(tzinfo=dateutil.tz.tzutc())

# Twitter Setup

# API setup.
apiConfigPath = '../../../conf/twitter.json'
# Read API settings.
twitterSettings = json.load(open(apiConfigPath))
# Set up twitter object.
twitter = Twython(twitterSettings['apiKey'], access_token=twitterSettings['accessToken'])
# Wait time. Ensure we don't run afoul of rate limits.
secondsToWait = 6

# Organization/SN setup.
allOrgs = json.load(open('../../conf/organizations.json'))
sn = allOrgs[org]

# Data directory.
consolidatedTweetsDirectory = '../../../data/twitter/consolidated/%s/' % org
if not os.path.exists(consolidatedTweetsDirectory):
    raise Exception('No consolidated tweets for organization "%s" (@%s).' % (org,sn))

# Data directory.
rawTweetsDirectory = '../../../data/twitter/raw/%s/' % org
# Get timestamp of current run.
timestampFilename = dt.datetime.now().strftime('%Y-%m-%dT%H-%M-%S') + '.json'

# Base search URL. Can just append an ID after this.
baseSearchUrl = 'https://twitter.com/search?f=realtime&q=from%%3A%s%%20max_id%%3A' % sn

### Find Starting Point

# Find all consolidated tweet files and get the most recent.
tweetFile = [f for f in os.listdir(consolidatedTweetsDirectory) if f[-5:] == '.json'][-1]

# Load that file with json.
consolidatedTweets = json.load(open(consolidatedTweetsDirectory+tweetFile))

# Find min ID.
minTweetId = None
for tweet in consolidatedTweets:
    minTweetId = min([tId for tId in [minTweetId, tweet['id']] if tId is not None])

# Status.
print 'Found minimum tweet ID for organization "%s" (@%s): %d' % (org,sn,minTweetId)
print 'Will start search at %d - 1' % minTweetId

### Scraping

# Container for tweets.
tweets = []

# Statistics.
totalTweets = 0
minTweetDate = None
maxTweetDate = None

# Loop terminates with break condition inside once we've found a date.
while True:
    # Decrement ID to avoid repeats.
    minTweetId -= 1

    # Make request and convert to a bs4 object.
    requestUrl = baseSearchUrl + str(minTweetId)
    searchRequest = requests.get(requestUrl)
    soup = BeautifulSoup(searchRequest.text)

    # Traverse all content items and extract tweet data.
    tweetContainers = soup.find_all('div', class_='tweet')

    # Traverse all tweet containers.
    for t in tweetContainers:
        # Find tweet ID for this tweet.
        tweetId = int(t.attrs['data-tweet-id'])

        # Get tweet; append.
        tweet = twitter.show_status(id=tweetId)
        tweets.append(tweet)

        # Get stats.
        minTweetId = minTweetId = min([tId for tId in [minTweetId, tweetId] if tId is not None])
        tweetDate = dateutil.parser.parse(tweet['created_at'])
        minTweetDate = min([tDate for tDate in [minTweetDate, tweetDate] if tDate is not None])
        maxTweetDate = max([tDate for tDate in [maxTweetDate, tweetDate] if tDate is not None])

        # Wait before making another request.
        time.sleep(secondsToWait)

    # Write out if we've hit the per-file limit or have gone back far enough.
    if len(tweets) >= tweetsPerFile or minTweetDate < until:
        # Stats.
        totalTweets += len(tweets)

        # Status.
        print 'Writing %d tweets to file %s' % (len(tweets), timestampFilename)

        # Output to file.
        filename = rawTweetsDirectory + timestampFilename
        json.dump(tweets, open(filename,'w'))

        # Reset relevant objects.
        timestampFilename = dt.datetime.now().strftime('%Y-%m-%dT%H-%M-%S') + '.json'
        tweets = []

        # Break if we've gone back far enough.
        if minTweetDate < until:
            break

Facebook Posts

We use Facebook's [Graph API](https://developers.facebook.com/docs/graph-api) to retrieve recent posts by each user. Facebook imposes no burdensome history length limit, so our code simply requests information in 75-post chunks until hitting a specified past date (set to September 1 when we gathered our data). See the full version in: [/src/fb/downloadRawFbPosts.py](https://github.com/micahlanier/social-news/blob/master/src/fb/downloadRawFbPosts.py).

"""

This script will pull recent Facebook posts from given usernames and dump them to a timestamped file.
If we pull posts over time, this will allow us to arrange pulls in order and consolidate them into a master file.

"""

##### Configuration

# List of organizations to fetch. Either a list or "all".
orgs = 'all'
# Upper limit of date.
maxDate = dt.datetime(2014,11,25,tzinfo=dateutil.tz.tzutc()) #dt.datetime.utcnow()
# Lower limit of date.
minDate = dt.datetime(2014, 8,31,tzinfo=dateutil.tz.tzutc())
# Facebook is not as explicit as Twitter about limiting. Set this as needed if you run into any issues.
postLimit = 75

##### Main Setup

# API setup.
apiConfigPath = '../../../conf/facebook.json'

# Data directory.
rawPostsDirectory = '../../../data/facebook/raw/'
# Get timestamp of current run.
timestampFilename = dt.datetime.now().strftime('%Y-%m-%dT%H-%M-%S') + '.json'

# Facebook Setup
# Read API settings.
fbSettings = json.load(open(apiConfigPath))

# Ensure maxDate comparable to minDate.
maxDate = maxDate.replace(tzinfo=dateutil.tz.tzutc())

# Keep a one-second time delta handy for upper-bound limits.
oneSecond = dt.timedelta(seconds=1)

# Sleep interval for transient errors.
sleepTimeSeconds = 4

# Organization setup.
allOrgs = json.load(open('../../conf/organizations.json'))
if (type(orgs) == type('str')):
    orgs = allOrgs.keys()
relevantOrgs = dict((k, allOrgs[k]) for k in orgs)

##### Facebook Helper Code

"""
Simple function for performing a Facebook GET request.
Does not check parameters, so be careful!
"""
def facebookGet(path, accessToken, parameters, version = '2.1'):
    fbUrl = 'https://graph.facebook.com/v' + version + '/'
    # Add request path.
    fbUrl += path.strip('/') + '/'
    # Add access token.
    fbUrl += '?access_token=' + accessToken
    # Add additional parameters.
    for pKey, pValue in parameters.iteritems():
        fbUrl += '&%s=%s' % (pKey, urllib.quote(str(pValue)))
    return requests.get(fbUrl)

##### Get Facebook Posts

# Iterate over organizations.
for org, orgData in relevantOrgs.iteritems():
    # Get account.
    acct = orgData['facebook']

    print 'Retrieving posts for organization "%s" (@%s).' % (org,acct)

    # Set up folder if it doesn't exist.
    acctDirectory = rawPostsDirectory + org + '/'
    if not os.path.exists(acctDirectory):
        os.makedirs(acctDirectory)

    # Set up container for posts.
    posts = []

    # Set up a loop date.
    loopMaxDate = maxDate

    # Set up objects hold our request.
    requestPath = acct + '/posts'
    requestParams = dict()
    requestParams['limit'] = postLimit

    # Don't bother with getting ALL comments/likes. We're only interested in summary statistics.
    # While it would be cool to analyze all of the people who comment/like, that's a PITA in the time we have.
    # By doing this, we need to specify a whitelist of all fields that we want.
    apiFields = [
        'comments.limit(1).summary(true)','likes.limit(1).summary(true)',
        'caption','created_time','description','from','id','link','message','name','shares','status_type','type'
    ]
    requestParams['fields'] = ','.join(apiFields)

    # Reload settings just in case we've changed access tokens and are running manually.
    fbSettings = json.load(open(apiConfigPath))

    # Loop until we have gone back to minDate.
    while loopMaxDate > minDate:
        # Compose a request with the latest max date.
        requestParams['until'] = str(loopMaxDate)
        # Make request; parse JSON.
        facebookRequest = facebookGet(requestPath, fbSettings['accessToken'], requestParams)
        jsonFacebookRequest = json.loads(facebookRequest.text)

        # Detect any errors.
        if 'error' in jsonFacebookRequest:
            if jsonFacebookRequest['error']['code'] == 2:
                # Tranient, "unexpected" error. Sleep for a sec and then continue.
                print 'Unexpected error. Sleeping %d seconds.' % sleepTimeSeconds
                time.sleep(2)
                continue
            else:
                print 'Encountered a major error. Breaking loop.'
                print jsonFacebookRequest['error']
                break

        # If no errors, we got data.
        currentPosts = jsonFacebookRequest['data']

        # Break if we didn't get anything; otherwise append and continue.
        if (not len(currentPosts)):
            break
        posts += currentPosts

        # Get date for next request. Go back one second, because for some reason they use inclusive logic.
        loopMaxDate = dateutil.parser.parse(currentPosts[-1]['created_time']) - oneSecond

    # Dump to file.
    filename = acctDirectory + timestampFilename
    json.dump(posts, open(filename,'w'), indent=0)

    # Get max/min dates retrieved.
    minDateRetrieved = loopMaxDate + oneSecond
    maxDateRetrieved = dateutil.parser.parse(posts[0]['created_time'])

User Data

The Twitter and Facebook APIs both provide easy access to user information. These scripts retrieve that information for display, and store it in the /data/users/ directory. See the full version in: [/src/users/user_statistics.ipynb](http://nbviewer.ipython.org/github/micahlanier/social-news/blob/master/src/users/user_statistics.ipynb).

##### Setup
# Get information about organizations.
orgs = json.load(open('../../conf/organizations.json'))
# Convert to screen names data structure.
screenNames = dict(
    (orgData['name'],{'twitter': orgData['twitter'], 'facebook': orgData['facebook'], 'org': org})
    for (org,orgData) in orgs.iteritems()
)
# Get Twitter settings from outside of the repository path.
twitterSettingsPath = '../../../conf/twitter.json'
twitterSettings = json.load(open(twitterSettingsPath))
# Get access token if we need one.
if ('accessToken' not in twitterSettings):
    twitter = Twython(twitterSettings['apiKey'], twitterSettings['apiSecret'], oauth_version=2)
    twitterSettings['accessToken'] = twitter.obtain_access_token()
    json.dump(open(twitterSettings,'w'))
# Set up our Twython object.
twitter = Twython(twitterSettings['apiKey'], access_token=twitterSettings['accessToken'])

##### Twitter
# Get user information.
twNames = []
twNameRevMap = {}
for name in screenNames:
    twNames.append(screenNames[name]['twitter'])
    twNameRevMap[screenNames[name]['twitter']] = name
stats = twitter.lookup_user(screen_name=','.join(twNames))
column_mapper = {
    'favourites_count' : 'favourites_count',
    'listed_count' : 'listed_count',
    'followers_count' : 'followers_count',
    'statuses_count' : 'statuses_count',
    'friends_count' : 'friends_count',
}
res = {}
for key in column_mapper:
    res[key] = {}
for st in stats:
    for key in column_mapper:
        res[key][twNameRevMap[st['screen_name'].lower()]] = st[column_mapper[key]]
df_twitter = pd.DataFrame(res)
# Save to data folder.
# Transpose and convert to dict.
orgStats = dict((screenNames[i]['org'],j) for (i,j) in df_twitter.transpose().to_dict().iteritems())
# Save.
json.dump(orgStats,open('../../data/users/twitter.json','w'),indent=4)

### Facebook
fbNameRevMap = {}
for name in screenNames:
    fbNameRevMap[screenNames[name]['facebook']] = name

fb_stats = {}
for name in screenNames:
    fb_stats[name] = json.load(urllib2.urlopen("http://graph.facebook.com/"+screenNames[name]['facebook']))
    fb_column_mapper = {
    'id' : 'id',
    'likes' : 'likes',
    'talking_about_count' : 'talking_about_count',
}
fb_res = {}

for key in fb_column_mapper:
    fb_res[key] = {}

for name in screenNames:
    for key in fb_column_mapper:
        fb_res[key][name] = fb_stats[name][fb_column_mapper[key]]
        
df_facebook = pd.DataFrame(fb_res)
# Save to data folder.
# Transpose and convert to dict.
orgStats = dict((screenNames[i]['org'],j) for (i,j) in df_facebook.transpose().to_dict().iteritems())
# Save.
json.dump(orgStats,open('../../data/users/facebook.json','w'),indent=4)

Twitter & Facebook Consolidation

The above scripts are designed to be run progressively in order to collect data as time goes on. Their data may overlap significantly, so it is necessary to consolidate all retrieved data into a "master" JSON file for each organization-network. In retrospect, a NoSQL solution like MongoDB would have been a logical choice for storing this information, and we would certainly use it if we had more time available (and if we opt to continue our analysis in the future). See the full versions of these files in: [/src/fb/consolidateRawFbPosts.py](https://github.com/micahlanier/social-news/blob/master/src/fb/consolidateRawFbPosts.py) and [/src/twitter/consolidateRawTweets.py](https://github.com/micahlanier/social-news/blob/master/src/twitter/consolidateRawTweets.py).

Twitter Consolidation

"""

This script will consolidate raw tweet files into one large JSON file (timestamped by runtime).

"""

##### Configuration

# Organizations to consolidate. Either a list or "all".
orgs = 'all'

##### Main Setup

# Data directory.
rawTweetsDirectory = '../../../data/twitter/raw/'
consolidatedTweetsDirectory = '../../../data/twitter/consolidated/'
# Get timestamp of current run.
timestampFilename = dt.datetime.now().strftime('%Y-%m-%dT%H-%M-%S')

# Organization setup.
allOrgs = json.load(open('../../conf/organizations.json'))
if (type(orgs) == type('str')):
    orgs = allOrgs.keys()
relevantOrgs = dict((k, allOrgs[k]) for k in orgs);

##### Get Tweets

# Iterate over organizations.
for org, orgData in relevantOrgs.iteritems():
    # Get SN.
    sn = orgData['twitter']

    # We're going to hold tweets in one dict.
    # If we end up with a lot of them we may need to be more careful here.
    consolidatedTweets = dict()

    # Keep track of tweets read; min/max date.
    tweetsRead = 0
    minTweetDate = None
    maxTweetDate = None

    # Get list of files. Remove junk in the process.
    rawTweetsFiles = [rtf for rtf in os.listdir(rawTweetsDirectory + org + '/') if rtf[-5:] == '.json']

    # Iterate over files.
    for rawTweetFile in rawTweetsFiles:
        # Read in JSON.
        filepath = rawTweetsDirectory + org + '/' + rawTweetFile
        rawTweets = json.load(open(filepath))
        # Iterate over tweets.
        for rawTweet in rawTweets:
            # Get tweet and use it as key for insertion into consolidatedTweets.
            tId = rawTweet['id']
            consolidatedTweets[tId] = rawTweet
            # Track tweet date info.
            currentTweetDate = dateutil.parser.parse(rawTweet['created_at'])
            minTweetDate = min([tDate for tDate in [minTweetDate, currentTweetDate] if tDate is not None])
            maxTweetDate = max([tDate for tDate in [maxTweetDate, currentTweetDate] if tDate is not None])
        # Keep track of tweets read.
        tweetsRead += len(rawTweets)

    # Dump to file.
    filename = consolidatedTweetsDirectory + '/' + org + '-' + timestampFilename + '.json'
    json.dump(consolidatedTweets.values(), open(filename,'w'))

Facebook Consolidation

"""

This script will consolidate raw post files into one large JSON file (timestamped by runtime).

"""

##### Configuration

# Edit things here to your liking. But please don't commit them needlessly.

# List of organizations to fetch. Either a list or "all".
orgs = 'all'

##### Main Setup

# Data directory.
rawPostsDirectory = '../../../data/facebook/raw/'
consolidatedPostsDirectory = '../../../data/facebook/consolidated/'
# Get timestamp of current run.
timestampFilename = dt.datetime.now().strftime('%Y-%m-%dT%H-%M-%S')

# Organization setup.
allOrgs = json.load(open('../../conf/organizations.json'))
if (type(orgs) == type('str')):
    orgs = allOrgs.keys()
relevantOrgs = dict((k, allOrgs[k]) for k in orgs)

##### Get Tweets

# Iterate over organizations.
for org, orgData in relevantOrgs.iteritems():
    # Get account.
    acct = orgData['facebook']

    # We're going to hold posts in one dict.
    # If we end up with a lot of them we may need to be more careful here.
    consolidatedPosts = dict()

    # Keep track of posts read; min/max date.
    postsRead = 0
    minPostDate = None
    maxPostDate = None

    # Get list of files. Remove junk in the process.
    rawPostsFiles = [rtf for rtf in os.listdir(rawPostsDirectory + org + '/') if rtf[-5:] == '.json']

    # Iterate over files.
    for rawPostFile in rawPostsFiles:
        # Read in JSON.
        filepath = rawPostsDirectory + org + '/' + rawPostFile
        rawPosts = json.load(open(filepath))
        # Iterate over tweets.
        for rawPost in rawPosts:
            # Get post and use it as key for insertion into consolidatedPosts.
            pId = rawPost['id']
            consolidatedPosts[pId] = rawPost
            # Track tweet date info.
            currentPostDate = dateutil.parser.parse(rawPost['created_time'])
            minPostDate = min([tDate for tDate in [minPostDate, currentPostDate] if tDate is not None])
            maxPostDate = max([tDate for tDate in [maxPostDate, currentPostDate] if tDate is not None])
        # Keep track of posts read.
        postsRead += len(rawPosts)

    # Dump to file.
    filename = consolidatedPostsDirectory + org + '-' + timestampFilename + '.json'
    json.dump(consolidatedPosts.values(), open(filename,'w'))

URL Retrieval

We were interested in determining where social media links actually *go*. That knowledge would enable us to (1) track overall performance of a link that is posted more than once (e.g., [@nytimes's football map tweets](https://twitter.com/search?q=from%3Anytimes%20football%20map&src=typd)) and (2) match Twitter and Facebook posts in order to determine relative effectiveness. This is obscured by the fact that 23 of the 25 news organizations that we examined use Bitly regularly for posting links on social media websites. But Bitly API calls cannot necessarily tell us where those links go: many redirect several more times (e.g., through [trib.al](http://trib.al/)) before reaching their final destinations. Thus, we were forced to determine link destinations on our own.

We encountered two problems accomplishing this. First, we did not want to inflate Bitly click counts by following those links directly. This was alleviated by making our requests with HTTP `HEAD` requests, which avoid doing so. And second, we found that some news websites would return errors if we made high-frequency requests to them. To get around this, our code traverses the URL redirection chain until encountering the final link, but does not follow it.

This code is designed to be executed from the command line in order to better handle parallelization and error recovery. See full versions of this code in [/src/urls/socialUrlUtils.py](https://github.com/micahlanier/social-news/blob/master/src/urls/socialUrlUtils.py) and [/src/urls/scrapeUrls.py](https://github.com/micahlanier/social-news/blob/master/src/urls/scrapeUrls.py). The former appears first; the latter, second.

"""

Utilities for detecting certain kinds of Facebook links and extracting links from text.

Uses URL-detection regular expression from:
    https://gist.github.com/gruber/8891611

"""

# Define pattern.
urlPattern = '(?i)\\b((?:https?:(?:/{1,3}|[a-z0-9%])|[a-z0-9.\\-]+[.](?:com|net|org|edu|gov|mil|aero|asia|biz|cat|coop|info|int|jobs|mobi|museum|name|post|pro|tel|travel|xxx|ac|ad|ae|af|ag|ai|al|am|an|ao|aq|ar|as|at|au|aw|ax|az|ba|bb|bd|be|bf|bg|bh|bi|bj|bm|bn|bo|br|bs|bt|bv|bw|by|bz|ca|cc|cd|cf|cg|ch|ci|ck|cl|cm|cn|co|cr|cs|cu|cv|cx|cy|cz|dd|de|dj|dk|dm|do|dz|ec|ee|eg|eh|er|es|et|eu|fi|fj|fk|fm|fo|fr|ga|gb|gd|ge|gf|gg|gh|gi|gl|gm|gn|gp|gq|gr|gs|gt|gu|gw|gy|hk|hm|hn|hr|ht|hu|id|ie|il|im|in|io|iq|ir|is|it|je|jm|jo|jp|ke|kg|kh|ki|km|kn|kp|kr|kw|ky|kz|la|lb|lc|li|lk|lr|ls|lt|lu|lv|ly|ma|mc|md|me|mg|mh|mk|ml|mm|mn|mo|mp|mq|mr|ms|mt|mu|mv|mw|mx|my|mz|na|nc|ne|nf|ng|ni|nl|no|np|nr|nu|nz|om|pa|pe|pf|pg|ph|pk|pl|pm|pn|pr|ps|pt|pw|py|qa|re|ro|rs|ru|rw|sa|sb|sc|sd|se|sg|sh|si|sj|Ja|sk|sl|sm|sn|so|sr|ss|st|su|sv|sx|sy|sz|tc|td|tf|tg|th|tj|tk|tl|tm|tn|to|tp|tr|tt|tv|tw|tz|ua|ug|uk|us|uy|uz|va|vc|ve|vg|vi|vn|vu|wf|ws|ye|yt|yu|za|zm|zw)/)(?:[^\\s()<>{}\\[\\]]+|\\([^\\s()]*?\\([^\\s()]+\\)[^\\s()]*?\\)|\\([^\\s]+?\\))+(?:\\([^\\s()]*?\\([^\\s()]+\\)[^\\s()]*?\\)|\\([^\\s]+?\\)|[^\\s`!()\\[\\]{};:\'".,<>?«»“”‘’])|(?:(?<!@)[a-z0-9]+(?:[.\\-][a-z0-9]+)*[.](?:com|net|org|edu|gov|mil|aero|asia|biz|cat|coop|info|int|jobs|mobi|museum|name|post|pro|tel|travel|xxx|ac|ad|ae|af|ag|ai|al|am|an|ao|aq|ar|as|at|au|aw|ax|az|ba|bb|bd|be|bf|bg|bh|bi|bj|bm|bn|bo|br|bs|bt|bv|bw|by|bz|ca|cc|cd|cf|cg|ch|ci|ck|cl|cm|cn|co|cr|cs|cu|cv|cx|cy|cz|dd|de|dj|dk|dm|do|dz|ec|ee|eg|eh|er|es|et|eu|fi|fj|fk|fm|fo|fr|ga|gb|gd|ge|gf|gg|gh|gi|gl|gm|gn|gp|gq|gr|gs|gt|gu|gw|gy|hk|hm|hn|hr|ht|hu|id|ie|il|im|in|io|iq|ir|is|it|je|jm|jo|jp|ke|kg|kh|ki|km|kn|kp|kr|kw|ky|kz|la|lb|lc|li|lk|lr|ls|lt|lu|lv|ly|ma|mc|md|me|mg|mh|mk|ml|mm|mn|mo|mp|mq|mr|ms|mt|mu|mv|mw|mx|my|mz|na|nc|ne|nf|ng|ni|nl|no|np|nr|nu|nz|om|pa|pe|pf|pg|ph|pk|pl|pm|pn|pr|ps|pt|pw|py|qa|re|ro|rs|ru|rw|sa|sb|sc|sd|se|sg|sh|si|sj|Ja|sk|sl|sm|sn|so|sr|ss|st|su|sv|sx|sy|sz|tc|td|tf|tg|th|tj|tk|tl|tm|tn|to|tp|tr|tt|tv|tw|tz|ua|ug|uk|us|uy|uz|va|vc|ve|vg|vi|vn|vu|wf|ws|ye|yt|yu|za|zm|zw)\\b/?(?!@)))'

"""
Detect if a URL is a link to a Facebook photo.
Examples:
    urlIsFacebookPhoto('https://www.facebook.com/bbcworldnews/photos/pcb.10153822033402588/10153822032512588/?type=1') == True
    urlIsFacebookPhoto('http://bbc.in/fgh543q') == False    
"""
def urlIsFacebookPhoto(url):
    return bool(re.match('^https?://(www\\.)?facebook.com/[a-zA-Z0-9]+/photos/.+$', url))

"""
Detect if a URL is a link to a Facebook video.
Examples:
    urlIsFacebookPhoto('https://www.facebook.com/video.php?v=10150471411394999') == True
    urlIsFacebookPhoto('http://bbc.in/fgh543q') == False    
"""
def urlIsFacebookVideo(url):
    return bool(re.match('^https?://(www\\.)?facebook.com/video.php\\?.+$', url))

"""
Return a list of all detected URLs, or an empty list if none.
"""
def urlsInText(t):
    return re.findall(urlPattern, t)

"""

This script scrapes URLs from posts and tweets. See /doc/URLs.txt for information on how it works and how it stores data between runs.
At a high level, it begins by injesting all specified posts/tweets, and identifies the links therein.
Any that have already been scraped are ignored. We may eventually need a system for fixing bad results.
Any that have not been scraped are called with the requests module. For successful calls, we log ALL URLs traversed (just in case we need them).
We also dump request text to a file named using an MD5 hash of the final URL minus any query string.

Note that this script contains little inline configuration. URL scraping takes a while because we need to add delays so as to not cause 54 errors.
The best solution is to parallelize execution. But parallelization in this script is a pain, so just do it by running the script several times simultaneously.

To run, execute with the following parameters:
    python scrapeUrls.py [network] [username] [delay] [scrapeContent]
Parameters:
    * organization: organization name for the given network
    * network: one of 'facebook' or 'twitter'
    * delay: delay in seconds to wait before retrieving subsequent URLs
    * scrapeContent: 1 to scrape content, 0 to stop before making any request to the final link destination

Scraping is complicated. Some hosts are fine with 1 or two seconds delay (e.g., nytimes). Others need 4+ seconds (cnn).
Setting delays for this is more art than science, so good luck.

"""

##### Configuration

# This script is going to take a long time to run, so we want to store URLs in chunks in case anything goes wrong.
incrementalUrlsToStore = 25

# Timeout for hanging URLs.
timeoutSeconds = 10.0

##### Main Setup

# Our own custom URL parsing code for URLs.
import socialUrlUtils

# Data directories.
dirPath_consolidated = '../../../data/%s/consolidated/'
filePath_urls = '../../../data/urls/%s/%s.json'
dirPath_content = '../../../data/content/%s/%s/'
dirPath_urlBackups = '../../../data/backup/urls/%s/'

# Current timestamp.
currTimestamp = dt.datetime.now().strftime('%Y-%m-%dT%H-%M-%S')

##### Configure Source

# Set network, screen name, and sleep time based on inputs in sys.argv.
org = str(sys.argv[1])
network = str(sys.argv[2])
sleepSeconds = float(sys.argv[3])
scrapeContent = True if int(sys.argv[4]) else 0

# Get organizational information.
allOrgs = json.load(open('../../conf/organizations.json'))

# Get information about the organization's URLs.
targetDomain = allOrgs[org]['urls']['domain']
bitlyUrl = allOrgs[org]['urls']['bitly']
stillTraversePattern = allOrgs[org]['urls']['stillTraverse'] if 'stillTraverse' in allOrgs[org]['urls'] else None

# Any other miscellanea.
orgName = allOrgs[org]['name']

##### Helper Functions

"""
Return a list of expanded URLs regardless of service. Don't do any cleaning of URLs.
"""
def associatedUrls(network,post):
    if network == 'facebook':
        links = []
        if 'link' in post:
            links += [post['link']]
        if 'message' in post:
            links += socialUrlUtils.urlsInText(post['message'])
        return links
    elif network == 'twitter':
        return [u['expanded_url'] for u in post['entities']['urls']]
    return []

"""
Traverse URLs until encountering a target domain.
We may want to do this if a site makes an effort to counter scraping or raises "connection reset" errors.

Returns two values: final request object, an error string, and listing of traversed URLs.
If it encounters targetDomain, the final request object will be None. Otherwise, it will contain the final request.
If any request errors, the error's string representation will be passed back.
"""
def traverseUrls(firstUrl, targetDomain, stillTraversePattern, bitlyUrl, scrape=False):
    # Keep track of history.
    hist = []
    # Other return objects to keep.
    errorText = None
    lastRequest = None

    # Keep looping until we no longer have a URL to traverse.
    # We may manually break the loop inside if we want to exit immediately.
    nextUrl = firstUrl
    while nextUrl is not None:
        # Append the URL to history.
        hist += [nextUrl]
        # Parse URL.
        urlParsed = urlparse(nextUrl)
        # If the destination url is our target and we're not scraping, return the list leading up to it.
        # We test a LOT of conditions here. Break and return if the following is true:
        #     The domain is indeed the "target" domain (e.g., nytimes.com).
        #     The domain is not part of a bitly link (handles cases like the on.wsj.com bitly domain).
        #    The domain does not match an internal redirection page (handles cases like http://www.theguardian.com/p/3p6jv/tw).
        if (
                not scrape
            and    targetDomain is not None and (urlParsed.netloc == targetDomain or urlParsed.netloc[-len(targetDomain)-1:] == '.' + targetDomain)
            and urlParsed.netloc != bitlyUrl
            and (stillTraversePattern is None or not re.match(stillTraversePattern,nextUrl))
        ):
            break
        # If it is not, request the URL without redirects.
        # Use a get request if scraping.
        try:
            if (not scrape) or urlParsed.netloc in [bitlyUrl,'bit.ly','trib.al']:
                singleRequest = requests.head(nextUrl, allow_redirects=False, timeout=timeoutSeconds)
            else:
                singleRequest = requests.get(nextUrl, allow_redirects=False, timeout=timeoutSeconds)
        except:
            # We can experience connection errors for various reasons.
            # If we encounter one, append the URL to the end and move on.
            e = sys.exc_info()[0]
            errorText = str(e)
            break
        # We have a request. Look for a location in the header. If none, we've reached the end and can return.
        if 'location' not in singleRequest.headers:
            # Store our request as the last request.
            lastRequest = singleRequest
            break
        # Set the next URL to traverse.
        nextUrl = singleRequest.headers['location']
        # To avoid loops, verify if address has been traversed before.
        if nextUrl in hist:
            errorText = 'Loop encountered'
            break
        
    # Return.
    return lastRequest, errorText, hist

##### Get URLs

# Statistics.
urlsSkipped = 0
urlsRetrieved = 0
fbUrlsRetrieved = 0
twUrlsRetrieved = 0
postsSansUrl = 0
urlsScraped = 0

# Validation.
if network not in ['facebook','twitter']:
    print 'Network "%s" invalid; skipping user %s.' % (network, org)
    quit()

# Look for existing URLs.
orgUrlsFilename = (filePath_urls % (network,org))
urlsFileExists = os.path.exists(orgUrlsFilename)
# Create a URL dict if we don't find it.
urls = dict()
if urlsFileExists:
    # URLs file exists. Load it.
    urls = json.load(open(orgUrlsFilename))

    # Because we're going to be modifying it, create a backup.
    backupFilename = (dirPath_urlBackups % network) + org + '-' + currTimestamp + '.json'
    shutil.copyfile(orgUrlsFilename, backupFilename)
    print 'Backed up URLs file to %s' % '/'.join(backupFilename.split('/')[-2:])

# Get posts. Use the latest file if more than one.
postsFilename = [f for f in os.listdir(dirPath_consolidated % network) if f.find(org) == 0][-1]
posts = json.load(open(dirPath_consolidated % network + '/' + postsFilename))

# Set up content directory path. Create if doesn't exist.
contentDirectory = dirPath_content % (network, org)
if not os.path.exists(contentDirectory):
    os.makedirs(contentDirectory)

# Traverse all posts, urls.
# enumerate() pattern is for debugging. It allows us to break after a certain number of posts if necessary.
# Can remove before release.
for (i, post) in enumerate(posts):
    # Get links and traverse.
    postUrls = associatedUrls(network,post)
    if len(postUrls) == 0:
        postsSansUrl += 1
    for url in postUrls:
        # Check to see if URL has ever been checked before. If so, note it and continue.
        # In practice, this is like breaking (most posts only have one link).
        # But we must continue in case there is another link in the post that is valid.
        if url in urls:
            urlsSkipped += 1
            continue

        # Container for URL data. How we fill it depends on scrapeContent.
        urlData = dict()
        # Handle base content cases so we don't need to set "defaults" everywhere.
        urlData['errorText'] = None
        urlData['history'] = []
        urlData['statusCode'] = None
        urlData['scraped'] = scrapeContent
        content = None

        # Make request. Use a traverseUrl() call to avoid triggering connection resets or contaminating Bitly observations.
        finalRequest, errorText, urlsTraversed = traverseUrls(url,targetDomain,stillTraversePattern,bitlyUrl,scrapeContent)

        # Set up history, status codes. Codes only available if we made a final request.
        urlData['history'] = urlsTraversed
        if finalRequest is not None:
            urlData['statusCode'] = finalRequest.status_code

        # Handle errors.
        if errorText is not None:
            urlData['errorText'] = errorText

        # Store MD5. Used for determining write location of content if we scraped it.
        urlMd5 = md5(url.encode('utf8')).hexdigest()
        urlData['md5'] = urlMd5

        # Save content if we have it.
        if finalRequest is not None and bool(finalRequest.content):
            with open(contentDirectory+urlMd5+'.html','w') as f:
                f.write(finalRequest.content)
            urlsScraped += 1

        # Parse final URL.
        urlParsed = urlparse(urlData['history'][-1])
        urlData['endUrl'] = {
            'url': urlData['history'][-1],
            'scheme': urlParsed.scheme,
            'netloc': urlParsed.netloc,
            'path': urlParsed.path,
            'query': urlParsed.query,
            'fragment': urlParsed.fragment
        }

        # Store in main URLs dictionary.
        urls[url] = urlData

        # Statistics.
        urlsRetrieved += 1
        if (network == 'twitter'):
            twUrlsRetrieved += 1
        else:
            fbUrlsRetrieved += 1

        # Dump URLs if we've gotten to that point.
        if urlsRetrieved % incrementalUrlsToStore == 0:
            print 'Storing %d URLs.' % incrementalUrlsToStore
            json.dump(urls, open(orgUrlsFilename,'w'))

        # Wait, if we want to.
        time.sleep(sleepSeconds)

# Store any last URLs for this user.
json.dump(urls, open(orgUrlsFilename,'w'))

Bitly Traffic Data

Finally, Bitly provides a [REST API](http://dev.bitly.com/) for retrieving link click statistics. We simply built a script to traverse every link and retrieve its aggregate click count.

Like URL retrieval code, this script is designed to be executed from the command line in order to better handle parallelization and error recovery. See the full version in: [/src/bitly/getBitlyTraffic.py](https://github.com/micahlanier/social-news/blob/master/src/bitly/getBitlyTraffic.py).

"""

This script fetches Bitly traffic for all such links contained in social media posts.
All URLs for which traffic has already been fetched will be ignored.

"""

##### Configuration

# Last date/time to fetch traffic. This should be several days in the past. There is currently no way to re-fetch data.
lastDate = dt.datetime(2014,11,24,8,tzinfo=dateutil.tz.tzutc())

# Sleep time between requests. Bitly is opaque with their API limits, so there is no easy way to determine to set this.
sleepSeconds = 0.1

# This script is going to take a long time to run, so we want to store URLs in chunks in case anything goes wrong.
incrementalUrlsToStore = 50

# Timeout for hanging URLs.
timeoutSeconds = 10.0

##### Main Setup

# Our own custom URL parsing code for Facebook.
import socialUrlUtils

# Data directories.
dirPath_consolidated = '../../../data/%s/consolidated/'
filePath_bitly = '../../../data/bitly/%s/%s.json'
dirPath_bitlyBackups = '../../../data/backup/bitly/%s/'

# Current timestamp.
currTimestamp = dt.datetime.now().strftime('%Y-%m-%dT%H-%M-%S')

##### Main Configuration

# Set network, screen name, and sleep time based on inputs in sys.argv.
org = str(sys.argv[1])
network = str(sys.argv[2])

# Get organizational information.
allOrgs = json.load(open('../../conf/organizations.json'))

# Get organization information for reference.
orgName = allOrgs[org]['name']
bitlyUrl = allOrgs[org]['urls']['bitly']
bitlyUrls = [bitlyUrl,'bit.ly']

# API setup.
bitlySettings = json.load(open('../../../conf/bitly.json'))
bitlyAccessToken = bitlySettings['accessToken']
bitlyClicksBaseUrl = 'https://api-ssl.bitly.com/v3/link/clicks?'

##### Get Traffic

# Statistics.
urlsSkipped = 0
urlsSansBitlyUrl = 0
bitlyStatsRetrieved = 0
recentPostsSkipped = 0

# Validation.
if network not in ['facebook','twitter']:
    print 'Network "%s" invalid; skipping %s.' % (network, org)
    quit()

# Look for existing Bitly data.
orgBitlyFilename = (filePath_bitly % (network,org))
bitlyFileExists = os.path.exists(orgBitlyFilename)
# Create a Bitly dict if we don't find it.
bitly = dict()
if bitlyFileExists:
    # Bitly file exists. Load it.
    bitly = json.load(open(orgBitlyFilename))

    # Because we're going to be modifying it, create a backup.
    backupFilename = (dirPath_bitlyBackups % network) + org + '-' + currTimestamp + '.json'
    shutil.copyfile(orgBitlyFilename, backupFilename)
    print 'Backed up bitly file to %s' % '/'.join(backupFilename.split('/')[-2:])

# Get posts (for URLs).
postFilename = [f for f in os.listdir(dirPath_consolidated % network) if f.startswith(org)][-1]
posts = json.load(open((dirPath_consolidated % network) + postFilename))

# Traverse posts and find URLs.
urls = set()
if network == 'facebook':
    for p in posts:
        # Verify date.
        if dateutil.parser.parse(p['created_time']) < lastDate:
            # Find URLs to add.
            if 'link' in p:
                urls.add(p['link'])
            if 'message' in p:
                for url in socialUrlUtils.urlsInText(p['message']):
                    urls.add(url)
        else:
            recentPostsSkipped += 1
elif network == 'twitter':
    for p in posts:
        # Verify date.
        if dateutil.parser.parse(p['created_at']) < lastDate:
            # Find URLs to add.
            for url in p['entities']['urls']:
                urls.add(url['expanded_url'])
        else:
            recentPostsSkipped += 1

# Traverse all URLs.
# enumerate() pattern is for debugging. It allows us to break after a certain number of URLs if necessary.
for (i, url) in enumerate(urls):
    # Parse URL to determine if we should read.
    urlInfo = urlparse(url)
    # Examine domain. Ensure a bitly link; else continue.
    if (urlInfo.netloc not in bitlyUrls):
        urlsSansBitlyUrl += 1
        continue
    # Any URL at this point is readable against bitly.

    # Ensure that we haven't already retrieved the URL.
    # Later: we will want to add capability to go back to certain dates later.
    if (url in bitly):
        urlsSkipped += 1
        continue

    # Okay, now we know that we want to retrieve URL stats.
    # Construct our API call URL.
    requestQsParams = {
        'access_token': bitlyAccessToken,
        'link': urllib.quote(url)
    }
    requestUrl = bitlyClicksBaseUrl + '&'.join([k+'='+v for (k,v) in requestQsParams.iteritems()])

    # Make our request.
    bitlyRequest = requests.get(requestUrl)
    # Convert to dict.
    bitlyRequestData = json.loads(bitlyRequest.text)
    # Append a timestamp for our request.
    bitlyRequestData['retrieved'] = bitlyRequest.headers['date']

    # Store in main bitly container.
    bitly[url] = bitlyRequestData

    # Update statistics.
    bitlyStatsRetrieved += 1

    # Dump data if we've gotten to that point.
    if bitlyStatsRetrieved % incrementalUrlsToStore == 0:
        json.dump(bitly, open(orgBitlyFilename,'w'))

    # Wait, if we want to.
    time.sleep(sleepSeconds)

# Store any last link stats for this user.
json.dump(bitly, open(orgBitlyFilename,'w'))
