==> book/3.1_Data Wrangling.ipynb <==
#Summary:

###The following code extracts URL data for twitter and facebook posts. Then, it finds the same articles that are posted both on twitter and facebook###

# Libraries.
import json
import os
import pandas as pd
import numpy as np
from urlparse import urlparse

# Imports.
import re

# Define pattern.
urlPattern = '(?i)\\b((?:https?:(?:/{1,3}|[a-z0-9%])|[a-z0-9.\\-]+[.](?:com|net|org|edu|gov|mil|aero|asia|biz|cat|coop|info|int|jobs|mobi|museum|name|post|pro|tel|travel|xxx|ac|ad|ae|af|ag|ai|al|am|an|ao|aq|ar|as|at|au|aw|ax|az|ba|bb|bd|be|bf|bg|bh|bi|bj|bm|bn|bo|br|bs|bt|bv|bw|by|bz|ca|cc|cd|cf|cg|ch|ci|ck|cl|cm|cn|co|cr|cs|cu|cv|cx|cy|cz|dd|de|dj|dk|dm|do|dz|ec|ee|eg|eh|er|es|et|eu|fi|fj|fk|fm|fo|fr|ga|gb|gd|ge|gf|gg|gh|gi|gl|gm|gn|gp|gq|gr|gs|gt|gu|gw|gy|hk|hm|hn|hr|ht|hu|id|ie|il|im|in|io|iq|ir|is|it|je|jm|jo|jp|ke|kg|kh|ki|km|kn|kp|kr|kw|ky|kz|la|lb|lc|li|lk|lr|ls|lt|lu|lv|ly|ma|mc|md|me|mg|mh|mk|ml|mm|mn|mo|mp|mq|mr|ms|mt|mu|mv|mw|mx|my|mz|na|nc|ne|nf|ng|ni|nl|no|np|nr|nu|nz|om|pa|pe|pf|pg|ph|pk|pl|pm|pn|pr|ps|pt|pw|py|qa|re|ro|rs|ru|rw|sa|sb|sc|sd|se|sg|sh|si|sj|Ja|sk|sl|sm|sn|so|sr|ss|st|su|sv|sx|sy|sz|tc|td|tf|tg|th|tj|tk|tl|tm|tn|to|tp|tr|tt|tv|tw|tz|ua|ug|uk|us|uy|uz|va|vc|ve|vg|vi|vn|vu|wf|ws|ye|yt|yu|za|zm|zw)/)(?:[^\\s()<>{}\\[\\]]+|\\([^\\s()]*?\\([^\\s()]+\\)[^\\s()]*?\\)|\\([^\\s]+?\\))+(?:\\([^\\s()]*?\\([^\\s()]+\\)[^\\s()]*?\\)|\\([^\\s]+?\\)|[^\\s`!()\\[\\]{};:\'".,<>?«»“”‘’])|(?:(?<!@)[a-z0-9]+(?:[.\\-][a-z0-9]+)*[.](?:com|net|org|edu|gov|mil|aero|asia|biz|cat|coop|info|int|jobs|mobi|museum|name|post|pro|tel|travel|xxx|ac|ad|ae|af|ag|ai|al|am|an|ao|aq|ar|as|at|au|aw|ax|az|ba|bb|bd|be|bf|bg|bh|bi|bj|bm|bn|bo|br|bs|bt|bv|bw|by|bz|ca|cc|cd|cf|cg|ch|ci|ck|cl|cm|cn|co|cr|cs|cu|cv|cx|cy|cz|dd|de|dj|dk|dm|do|dz|ec|ee|eg|eh|er|es|et|eu|fi|fj|fk|fm|fo|fr|ga|gb|gd|ge|gf|gg|gh|gi|gl|gm|gn|gp|gq|gr|gs|gt|gu|gw|gy|hk|hm|hn|hr|ht|hu|id|ie|il|im|in|io|iq|ir|is|it|je|jm|jo|jp|ke|kg|kh|ki|km|kn|kp|kr|kw|ky|kz|la|lb|lc|li|lk|lr|ls|lt|lu|lv|ly|ma|mc|md|me|mg|mh|mk|ml|mm|mn|mo|mp|mq|mr|ms|mt|mu|mv|mw|mx|my|mz|na|nc|ne|nf|ng|ni|nl|no|np|nr|nu|nz|om|pa|pe|pf|pg|ph|pk|pl|pm|pn|pr|ps|pt|pw|py|qa|re|ro|rs|ru|rw|sa|sb|sc|sd|se|sg|sh|si|sj|Ja|sk|sl|sm|sn|so|sr|ss|st|su|sv|sx|sy|sz|tc|td|tf|tg|th|tj|tk|tl|tm|tn|to|tp|tr|tt|tv|tw|tz|ua|ug|uk|us|uy|uz|va|vc|ve|vg|vi|vn|vu|wf|ws|ye|yt|yu|za|zm|zw)\\b/?(?!@)))'

"""
Detect if a URL is a link to a Facebook photo.

Examples:
	urlIsFacebookPhoto('https://www.facebook.com/bbcworldnews/photos/pcb.10153822033402588/10153822032512588/?type=1') == True
	urlIsFacebookPhoto('http://bbc.in/fgh543q') == False	
"""
def urlIsFacebookPhoto(url):
	return bool(re.match('^https?://(www\\.)?facebook.com/[a-zA-Z0-9]+/photos/.+$', url))

"""
Detect if a URL is a link to a Facebook video.

Examples:
	urlIsFacebookPhoto('https://www.facebook.com/video.php?v=10150471411394999') == True
	urlIsFacebookPhoto('http://bbc.in/fgh543q') == False	
"""
def urlIsFacebookVideo(url):
	return bool(re.match('^https?://(www\\.)?facebook.com/video.php\\?.+$', url))

"""
TODO
"""
def urlsInText(t):
	# Perform search.
	urlMatches = re.search(urlPattern, t)
	if urlMatches:
		return [g for g in urlMatches.groups()]
	return []

def urlIsFacebookLink(url):
    return bool(re.match('^https?://(www\\.)?facebook.com.+$', url))

#Setup File Paths#

sources = ['twitter','facebook']

print "The current working directory is", os.getcwd()

pathConsolidated = os.getcwd()+'/Social News/data/3. urls/%s/'
pathExtractedUrl = os.getcwd()+'/Social News/data/4. extractedUrl/'
pathMergedUrl = os.getcwd()+'/Social News/data/5. mergedUrls/'

# Get all files to traverse. Assume that we can get everything in consolidated folder.
postFiles = []
for source in sources:
    sourcePath = pathConsolidated % source
    postFiles += [(source,p[:p.find('-')],sourcePath + p) for p in os.listdir(sourcePath) if p[-5:] == '.json']


# Uncomment to see all files read in
#postFiles

# Extract Organization Names
organizationNames = [x[1].split(".")[0] for x in postFiles]
organizationNames = set(organizationNames)
organizationNames = list(organizationNames)
organizationNames.sort()
len(organizationNames)

#Data Wrangling#

columns = ['NewsOrganization', 'PathUrl','matchUrl', 'FullUrl', 'ShortenedUrl', 'HistoryUrls']

columns_tw = ["TW_" + name for name in columns]
columns_fb = ["FB_" + name for name in columns]

"""
Function used to extract urls from the json objects into a dataframe
param[in] filenames: folder+filenames of the json objects
param[in] columnNames: column names for the resulting dataframe
param[in] socialMedia: 'twitter' or 'facebook'
param[in] destFolder: destination folder to save output csvs
"""
def extractUrltoDF(filenames, columnNames, socialMedia, destFolder):
    rowCount = 0
    facebookLinks = 0 # fb photo + video links
    facebookCount = 0 # all fb links
    facebookDict = {}
    
    print columnNames
    # Loop through each organization
    for organization in filenames:
        resultDF = pd.DataFrame(columns=columnNames) 
        if organization[0] == socialMedia:
            print organization[1]
            #print organization[2]
            data = json.load(open(organization[2]))

            #Iterate over dictionary to construct Dataframe
            for key, value in data.items():
                #print key
                if urlIsFacebookLink(key) and not (urlIsFacebookPhoto(key) or urlIsFacebookVideo(key)):
                    facebookCount += 1
                    facebookDict[key] = organization[1]
                if urlIsFacebookPhoto(key) or urlIsFacebookVideo(key):
                    facebookLinks += 1
                else:
                    newOrganization = value['endUrl']['netloc']
                    pathUrl = value['endUrl']['path']
                    matchUrl = newOrganization+pathUrl
                    fullUrl = value['endUrl']['url']
                    shortenedUrl = key
                    historyUrl = ""
                    for hUrl in value['history']:
                        historyUrl = historyUrl+"???"+hUrl
                    #Save results
                    resultDF.loc[rowCount] = [newOrganization, pathUrl, matchUrl, fullUrl, shortenedUrl, historyUrl]
                    rowCount += 1
            filename = organization[1].split('.')[0]
            resultDF.to_csv(destFolder+filename+'_'+socialMedia+'.csv', encoding='utf-8')
            del resultDF
    print "Number of rows: %d" % rowCount
    print "Number of fb photo + video: %d" % facebookLinks
    print "Number of fb links: %d" % facebookCount
    return rowCount, facebookDict

fbDF, fbArray = extractUrltoDF([postFiles[28]], columns_fb, 'facebook', pathExtractedUrl)

#Twitter Data#

#NYTimes example
#twitterDF = extractUrltoDF([postFiles[15]], columns_tw, 'twitter', pathExtractedUrl)

#Run all
twitterDF = extractUrltoDF(postFiles, columns_tw, 'twitter', pathExtractedUrl)

#Facebook Data#

#NYTimes example
#fbDF = extractUrltoDF([postFiles[40]], columns_fb, 'facebook', pathExtractedUrl)

#Run all
fbDF, notCapturedfbLinks = extractUrltoDF(postFiles, columns_fb, 'facebook', pathExtractedUrl)

#Merge Twitter and Facebook URLs#

"""
Function merges twitter and facebook article URLs
param[in] inputFolder: folder+filenames of the extracted url csv files
param[in] outputFolder: destination to save the output csv files
param[in] orgNames: names of the news organizations used to read and save csv files
"""
def mergeUrlFiles(inputFolder, outputFolder, orgNames):
    
    summaryDF = pd.DataFrame(columns=['newsOrg', 'tw_total', 'tw_unique', 'tw_repeat', 
                                      'fb_total', 'fb_unique', 'fb_repeat', 'merged_unique']) 
    rowCount = 0
    for orgName in orgNames:
        print orgName
        # Read in fb and tw url files and drop duplicates
        fb_DF = pd.read_csv( inputFolder+orgName+'_facebook.csv', index_col=0)
        fb_DF_unique = fb_DF.drop_duplicates(cols='FB_matchUrl', take_last=False)
        tw_DF = pd.read_csv( inputFolder+orgName+'_twitter.csv', index_col=0)
        tw_DF_unique = tw_DF.drop_duplicates(cols='TW_matchUrl', take_last=False)
        
        # Calculate number of duplicates
        fb_series = fb_DF['FB_matchUrl'].value_counts()
        fb_url_count = fb_series.to_frame('Counts')
        fb_url_count.to_csv(outputFolder+'DupStat/dup_count_'+orgName+'_fb.csv', encoding='utf-8')
        tw_series = tw_DF['TW_matchUrl'].value_counts()
        tw_url_count = tw_series.to_frame('Counts')
        tw_url_count.to_csv(outputFolder+'DupStat/dup_count_'+orgName+'_tw.csv', encoding='utf-8')
        
        # Merge both twitter and facebook data
        merged = pd.merge(tw_DF, fb_DF, left_on='TW_matchUrl', right_on='FB_matchUrl', how="inner")
        merged.to_csv(outputFolder+orgName+'_merged.csv', encoding='utf-8')
        merged_unique = merged.drop_duplicates(cols='TW_matchUrl', take_last=False)
        
        #Save results
        summaryDF.loc[rowCount] = [orgName, tw_DF.shape[0],tw_DF_unique.shape[0],tw_DF.shape[0]-tw_DF_unique.shape[0],
                                   fb_DF.shape[0],fb_DF_unique.shape[0],fb_DF.shape[0]-fb_DF_unique.shape[0],
                                   merged_unique.shape[0]]
        rowCount += 1
    return summaryDF
    

summaryStat = mergeUrlFiles(pathExtractedUrl, pathMergedUrl, organizationNames)
