==> book/3.2_Data Merging.ipynb <==
#Summary:

###The following code extracts social metrics data such as Facebook, Twitter, Bitly, Sentiment Social Media Data and merges them with matched Url Data###


# Libraries.
import json
import os
import pandas as pd
import numpy as np
from urlparse import urlparse

# Imports.
import re

# Define pattern.
urlPattern = '(?i)\\b((?:https?:(?:/{1,3}|[a-z0-9%])|[a-z0-9.\\-]+[.](?:com|net|org|edu|gov|mil|aero|asia|biz|cat|coop|info|int|jobs|mobi|museum|name|post|pro|tel|travel|xxx|ac|ad|ae|af|ag|ai|al|am|an|ao|aq|ar|as|at|au|aw|ax|az|ba|bb|bd|be|bf|bg|bh|bi|bj|bm|bn|bo|br|bs|bt|bv|bw|by|bz|ca|cc|cd|cf|cg|ch|ci|ck|cl|cm|cn|co|cr|cs|cu|cv|cx|cy|cz|dd|de|dj|dk|dm|do|dz|ec|ee|eg|eh|er|es|et|eu|fi|fj|fk|fm|fo|fr|ga|gb|gd|ge|gf|gg|gh|gi|gl|gm|gn|gp|gq|gr|gs|gt|gu|gw|gy|hk|hm|hn|hr|ht|hu|id|ie|il|im|in|io|iq|ir|is|it|je|jm|jo|jp|ke|kg|kh|ki|km|kn|kp|kr|kw|ky|kz|la|lb|lc|li|lk|lr|ls|lt|lu|lv|ly|ma|mc|md|me|mg|mh|mk|ml|mm|mn|mo|mp|mq|mr|ms|mt|mu|mv|mw|mx|my|mz|na|nc|ne|nf|ng|ni|nl|no|np|nr|nu|nz|om|pa|pe|pf|pg|ph|pk|pl|pm|pn|pr|ps|pt|pw|py|qa|re|ro|rs|ru|rw|sa|sb|sc|sd|se|sg|sh|si|sj|Ja|sk|sl|sm|sn|so|sr|ss|st|su|sv|sx|sy|sz|tc|td|tf|tg|th|tj|tk|tl|tm|tn|to|tp|tr|tt|tv|tw|tz|ua|ug|uk|us|uy|uz|va|vc|ve|vg|vi|vn|vu|wf|ws|ye|yt|yu|za|zm|zw)/)(?:[^\\s()<>{}\\[\\]]+|\\([^\\s()]*?\\([^\\s()]+\\)[^\\s()]*?\\)|\\([^\\s]+?\\))+(?:\\([^\\s()]*?\\([^\\s()]+\\)[^\\s()]*?\\)|\\([^\\s]+?\\)|[^\\s`!()\\[\\]{};:\'".,<>?«»“”‘’])|(?:(?<!@)[a-z0-9]+(?:[.\\-][a-z0-9]+)*[.](?:com|net|org|edu|gov|mil|aero|asia|biz|cat|coop|info|int|jobs|mobi|museum|name|post|pro|tel|travel|xxx|ac|ad|ae|af|ag|ai|al|am|an|ao|aq|ar|as|at|au|aw|ax|az|ba|bb|bd|be|bf|bg|bh|bi|bj|bm|bn|bo|br|bs|bt|bv|bw|by|bz|ca|cc|cd|cf|cg|ch|ci|ck|cl|cm|cn|co|cr|cs|cu|cv|cx|cy|cz|dd|de|dj|dk|dm|do|dz|ec|ee|eg|eh|er|es|et|eu|fi|fj|fk|fm|fo|fr|ga|gb|gd|ge|gf|gg|gh|gi|gl|gm|gn|gp|gq|gr|gs|gt|gu|gw|gy|hk|hm|hn|hr|ht|hu|id|ie|il|im|in|io|iq|ir|is|it|je|jm|jo|jp|ke|kg|kh|ki|km|kn|kp|kr|kw|ky|kz|la|lb|lc|li|lk|lr|ls|lt|lu|lv|ly|ma|mc|md|me|mg|mh|mk|ml|mm|mn|mo|mp|mq|mr|ms|mt|mu|mv|mw|mx|my|mz|na|nc|ne|nf|ng|ni|nl|no|np|nr|nu|nz|om|pa|pe|pf|pg|ph|pk|pl|pm|pn|pr|ps|pt|pw|py|qa|re|ro|rs|ru|rw|sa|sb|sc|sd|se|sg|sh|si|sj|Ja|sk|sl|sm|sn|so|sr|ss|st|su|sv|sx|sy|sz|tc|td|tf|tg|th|tj|tk|tl|tm|tn|to|tp|tr|tt|tv|tw|tz|ua|ug|uk|us|uy|uz|va|vc|ve|vg|vi|vn|vu|wf|ws|ye|yt|yu|za|zm|zw)\\b/?(?!@)))'

"""
Detect if a URL is a link to a Facebook photo.

Examples:
	urlIsFacebookPhoto('https://www.facebook.com/bbcworldnews/photos/pcb.10153822033402588/10153822032512588/?type=1') == True
	urlIsFacebookPhoto('http://bbc.in/fgh543q') == False	
"""
def urlIsFacebookPhoto(url):
	return bool(re.match('^https?://(www\\.)?facebook.com/[a-zA-Z0-9]+/photos/.+$', url))

"""
Detect if a URL is a link to a Facebook video.

Examples:
	urlIsFacebookPhoto('https://www.facebook.com/video.php?v=10150471411394999') == True
	urlIsFacebookPhoto('http://bbc.in/fgh543q') == False	
"""
def urlIsFacebookVideo(url):
	return bool(re.match('^https?://(www\\.)?facebook.com/video.php\\?.+$', url))

"""
TODO
"""
def urlsInText(t):
	# Perform search.
	urlMatches = re.search(urlPattern, t)
	if urlMatches:
		return [g for g in urlMatches.groups()]
	return []

def urlIsFacebookLink(url):
    return bool(re.match('^https?://(www\\.)?facebook.com.+$', url))

sources = ['twitter','facebook']

address = ['1. twitter', '2. facebook']

# File paths needed for data processing
pathConsolidated = os.getcwd()+'/Social News/data/%s/consolidated/'
pathMergedUrls = os.getcwd()+'/Social News/data/5. mergedUrls/'
pathSocialMediaMetrics = os.getcwd()+'/Social News/data/6. socialMetrics/'
pathBitly = os.getcwd()+'/Social News/data/7. bitly/'
pathSentiment = os.getcwd()+'/Social News/data/9. sentiment/'
pathFinalTable = os.getcwd()+'/Social News/data/8. finalTable/'

# Get all files to traverse. Assume that we can get everything in consolidated folder.
postFiles = []
for source in address:
    sourcePath = pathConsolidated % source
    postFiles += [(source,p[:p.find('-')],sourcePath + p) for p in os.listdir(sourcePath) if p[-5:] == '.json']

# Uncomment to see all files read in
#postFiles

# Extract Organization Names
organizationNames = [x[1].split(".")[0] for x in postFiles]
organizationNames = set(organizationNames)
organizationNames = list(organizationNames)
organizationNames.sort()
len(organizationNames)

#1. Extract Social Metrics Data#

"""
Function extracts different social metrics from RAW twitter and facebook posts JSON files depending on the social media medium
Twitter social metrics: twitter ID, twitter article URL, favorite, retweet, twitter text, twitter post time
Facebook social metrics: facebook ID, facebook article URL, like, share, facebook text, facebook post time
param[in] filenames: folder+filenames of the extracted url csv files
param[in] destFolder: destination to save the output csv files
"""
def extractSocialMetrics(filenames, destFolder):
    tw_columns = ['tw_id','tw_expandedUrl','tw_favorites', 'tw_retweet', 'tw_text', 'tw_postTime']
    fb_columns = ['fb_id','fb_expandedUrl','fb_likes', 'fb_share', 'fb_text', 'fb_postTime']
    rowCount = 0
    # Loop through each news organization to compile social metric statistics
    for organization in filenames:
        socialMedia = organization[0].split(' ')[-1]
        print organization[1]+" "+socialMedia
        if socialMedia == 'twitter':
            columnNames = tw_columns
        else:
            columnNames = fb_columns
        # Create DF with column names based on twitter or facebook
        resultDF = pd.DataFrame(columns=columnNames)
        
        # Load fb/tw json objects
        data = json.load(open(organization[2]))
        for article in data:
            # Data common to Twitter and Facebook.
            postId = article['id']
            #Twitter data
            if socialMedia == 'twitter':
                #Account for tweets from account holder or retweeting another tweet
                if 'retweeted_status' in article and 'favorite_count' in article['retweeted_status']:
                    favorites = article['retweeted_status']['favorite_count']
                elif 'favorite_count' in article:
                    favorites = article['favorite_count']
                else:
                    favorites = np.nan
                retweet = article['retweet_count']
                postTime = article['created_at']
                text = article['text']
                if len(article['entities']['urls']) == 0:
                    expandedUrl = np.nan
                else:
                    expandedUrl = article['entities']['urls'][0]['expanded_url']
            #Facebook data
            else:
                if 'link' in article and (urlIsFacebookPhoto(article['link']) or urlIsFacebookVideo(article['link'])):
                    #expandedUrl = urlsInText(article['message'])
                    if 'message' in article:
                        if len(urlsInText(article['message'])) > 0:
                            expandedUrl = urlsInText(article['message'])[0]
                        else:
                            expandedUrl = np.nan
                    else:
                        expandedUrl = np.nan
                elif 'link' in article:
                    expandedUrl = article['link']
                else:
                    expandedUrl = np.nan
                if 'likes' in article and 'summary' in article['likes'] and 'total_count' in article['likes']['summary']:
                    favorites = article['likes']['summary']['total_count']
                else:
                    favorites = np.nan
                if 'shares' in article and 'count' in article['shares']:
                    retweet = article['shares']['count']
                else:
                    retweet = np.nan
                postTime = article['created_time']
                if 'message' in article:
                    text = article['message']
                elif 'description' in article:
                    text = article['description']
                else:
                    text = np.nan
            resultDF.loc[rowCount] = [postId, expandedUrl, favorites, retweet, text, postTime]
            rowCount += 1
        #Save DF to csv file
        filename = organization[1].split('.')[0]
        resultDF.to_csv(destFolder+filename+'_'+socialMedia+'.csv', encoding='utf-8')

# NYTimes Example
#extractSocialMetrics([postFiles[15], postFiles[40]], tw_columns, pathSocialMediaMetrics)

# BBC Example
#extractSocialMetrics([postFiles[3], postFiles[28]], pathSocialMediaMetrics)

# Extract Social Metrics from all news organizations
extractSocialMetrics(postFiles, pathSocialMediaMetrics)

#Load in Bitly Data#

"""
Function extracts different bitly link counts from bitly json objects
param[in] bitlyInputFolder: folder+filenames of the bitly json objects
param[in] destFolder: destination to save the output csv files
param[in] orgNames: list of organization names used for looping through the files
param[in] socialMedia: 'twitter' or 'facebook'
"""
def extractBitly(bitlyInputFolder, destFolder, orgNames, socialMedia):
    # Loop through each orgagnization
    for orgName in orgNames:
        print orgName
        rowCount = 0
        # Use different column names depending on different social media medium
        if socialMedia == 'twitter':
            bitlyDF = pd.DataFrame(columns=['tw_url', 'tw_link_clicks', 'tw_timeRetrieved'])
        else:
            bitlyDF = pd.DataFrame(columns=['fb_url', 'fb_link_clicks', 'fb_timeRetrieved'])
        
        bitlyJson = json.load(open(bitlyInputFolder+orgName+'.json'))
        
        # 
        for key, value in bitlyJson.items():
            # Make sure each bitly json object does not have an error
            if ('data' in value) and (value['status_txt'] != 'NOT_FOUND' and value['status_txt'] != 'UNKNOWN_ERROR'):
                links_clicks = value['data']['link_clicks']
            # Skip bitly links if corrupted
            else:
                links_clicks = np.nan
            value['retrieved']
            #Save results
            bitlyDF.loc[rowCount] = [key, links_clicks, value['retrieved']]
            rowCount += 1
        print bitlyDF.shape
        bitlyDF.to_csv(destFolder+orgName+'_'+socialMedia+'.csv', encoding='utf-8')
    return bitlyDF

# Example
#extractBitly(pathBitly+'twitter/', pathBitly+'1. extractedBitly/', ['nbcnews'], 'twitter').head()

# Extract bitly links from twitter example
#extractBitly(pathBitly+'twitter/', pathBitly+'1. extractedBitly/', organizationNames, 'twitter').head()

# Extract bitly links from facebook example
#extractBitly(pathBitly+'facebook/', pathBitly+'1. extractedBitly/', organizationNames, 'facebook').head()

#Load in Sentiment Data#

"""
Function extracts sentiment data from sentiment analysis
For both Facebook and Twitter: social media ID, post's classified sentiment, negative sentiment raw score, positive sentiment raw score
param[in] sentimentInputFolder: folder+filenames of the sentiment data files
param[in] destFolder: destination to save the output csv files
param[in] orgNames: list of organization names used for looping through the files
param[in] socialMedia: 'twitter' or 'facebook'
"""
def extractSentiment(sentimentInputFolder, destFolder, orgNames, socialMedia):
    # Extract sentiment data for each news organization
    for orgName in orgNames:
        print orgName
        rowCount = 0
        cols = ['%s_id', '%s_sentiment_class', '%s_sentiment_negative', '%s_sentiment_positive']
        sentDF = pd.DataFrame(columns=[c % ('tw' if socialMedia == 'twitter' else 'fb') for c in cols])
        
        sentJson = json.load(open(sentimentInputFolder+orgName+'.json'))
        
        for key, value in sentJson.items():
            #print key
            #print value
            if value is not None:
                if socialMedia == 'twitter':
                    sent_class = value['class']
                    sent_neg = value['meanScoreNegSig']
                    sent_pos = value['meanScorePosSig']
                else:
                    sent_class = value['message']['class']
                    sent_neg = value['message']['meanScoreNegSig']
                    sent_pos = value['message']['meanScorePosSig']
            else:
                sent_class = sent_neg = sent_pos = None
            #Save results
            sentDF.loc[rowCount] = [key, sent_class, sent_neg, sent_pos]
            rowCount += 1
        print sentDF.shape
        sentDF.to_csv(destFolder+orgName+'_'+socialMedia+'.csv', encoding='utf-8')
    
    return sentDF    

# Extract sentiment from twitter
#extractSentiment(pathSentiment+'twitter/', pathSentiment+'1. extractedSentiment/', organizationNames, 'twitter').head()

# Extract sentiment from facebook
#extractSentiment(pathSentiment+'facebook/', pathSentiment+'1. extractedSentiment/', organizationNames, 'facebook').head()

#2. Merge the Facebook, Twitter, Bitly, Sentiment Social Media Data with Url Data

"""
Function uses mergedUrls files to merge with social metrics data
param[in] socialMedia: 'twitter' or 'facebook'
param[in] socialMedia_DF: data frame containing social media metrics data
param[in] socialMedia_mergedUrls_DF: data containing merged urls from twitter and facebook data
param[in] destFolder: destination to save the output intermediate csv files
param[in] orgName: python list of news organizations
"""
def sumSocialMetrics(socialMedia, socialMedia_DF, socialMedia_mergedUrls_DF, destFolder, orgName):
    # Twitter
    if socialMedia == 'twitter':
        # Drop duplicates in mergedUrls files
        social_w_Urls = socialMedia_DF.merge(socialMedia_mergedUrls_DF, left_on='tw_expandedUrl', right_on='TW_ShortenedUrl', how="inner")
        social_w_Urls = social_w_Urls.drop_duplicates(cols=['tw_expandedUrl', 'tw_postTime'] , take_last=True)
        social_w_Urls.to_csv(destFolder+'1. Intermediates/PreSum/'+orgName+'_PreSumSocialwUrls_twitter.csv', encoding='utf-8')
        # Account for posts that are reposted. Sum up the social metrics and keep count how many times the post is reposted
        social_w_Urls = social_w_Urls.groupby(['tw_expandedUrl']).agg([np.sum, len])
        social_w_Urls.to_csv(destFolder+'1. Intermediates/'+orgName+'_SocialwUrls_twitter.csv', encoding='utf-8')
    
        # Merge with social metrics data files
        DF_summed = socialMedia_DF.merge(social_w_Urls[['tw_favorites', 'tw_retweet', 'tw_text', 'tw_postTime']], 
                                   left_on='tw_expandedUrl', right_index=True , how="inner")
        DF_summed.columns = ['tw_id','tw_expandedUrl', 'tw_favorites', 'tw_retweet', 'tw_text', 'tw_postTime',
                                 'tw_favorites_Sum', 'tw_favorities_Count', 'tw_retweet_Sum', 'tw_retweet_Count',
                                 'tw_text_Sum', 'tw_text_Count', 'tw_text_Sum', 'tw_article_appearance_Count']
        DF_summed = DF_summed.drop(['tw_favorities_Count', 'tw_retweet_Count', 'tw_text_Count'], axis=1)
        DF_summed = DF_summed.drop_duplicates(cols='tw_expandedUrl', take_last=True)
        DF_summed.to_csv(destFolder+'1. Intermediates/'+orgName+'_SocialwUrls_twitter_merged.csv', encoding='utf-8')
    # Facebook
    else:
        # Drop duplicates in mergedUrls files
        social_w_Urls = socialMedia_DF.merge(socialMedia_mergedUrls_DF, left_on='fb_expandedUrl', right_on='FB_ShortenedUrl', how="inner")
        social_w_Urls = social_w_Urls.drop_duplicates(cols=['fb_expandedUrl', 'fb_postTime'] , take_last=True)
        social_w_Urls.to_csv(destFolder+'1. Intermediates/PreSum/'+orgName+'_PreSumSocialwUrls_facebook.csv', encoding='utf-8')
        # Account for posts that are reposted. Sum up the social metrics and keep count how many times the post is reposted
        social_w_Urls = social_w_Urls.groupby(['fb_expandedUrl']).agg([np.sum, len])
        social_w_Urls.to_csv(destFolder+'1. Intermediates/'+orgName+'_SocialwUrls_facebook.csv', encoding='utf-8')
    
        # Merge with social metrics data files
        DF_summed = socialMedia_DF.merge(social_w_Urls[['fb_likes', 'fb_share', 'fb_text', 'fb_postTime']], 
                                   left_on='fb_expandedUrl', right_index=True , how="inner")
        DF_summed.columns = ['fb_id','fb_expandedUrl', 'fb_likes', 'fb_share', 'fb_text', 'fb_postTime',
                                 'fb_likes_Sum', 'fb_likes_Count', 'fb_share_Sum', 'fb_share_Count',
                                 'fb_text_Sum', 'fb_text_Count', 'fb_text_Sum', 'fb_article_appearance_Count']
        DF_summed = DF_summed.drop(['fb_likes_Count', 'fb_share_Count', 'fb_text_Count'], axis=1)
        DF_summed = DF_summed.drop_duplicates(cols='fb_expandedUrl', take_last=True)
        DF_summed.to_csv(destFolder+'1. Intermediates/'+orgName+'_SocialwUrls_facebook_merged.csv', encoding='utf-8')
    return DF_summed

"""
Function used to create the final data table used for the analysis
Finaltable includes, matched url, social metrics, bitly data, and sentiment data
param[in] mergedUrlsInputFolder: folder containing twitter and facebook merged URLs data
param[in] socialMetricsInputFolder: folder containing twitter and facebook social metrics data
param[in] bitlyInputFolder: folder containing bitly data
param[in] sentimentInputFolder: folder containing sentiment data
param[in] orgNames: python list of news organizations
param[in] destFolder: destination to save the output final table csv files
"""
def mergeSocialMetricsData(mergedUrlsInputFolder, socialMetricsInputFolder, bitlyInputFolder, sentimentInputFolder, orgNames, destFolder):
    rowCount = 0
    facebookLinks = 0 # fb photo + video links
    facebookCount = 0 # all fb links
    facebookDict = {}
    tw_repostCounts = 0 # twitter repost counter
    fb_repostCounts = 0 # facebook repost counter
    
    for orgName in orgNames:
        print orgName
        # Read in social metrics data
        fb_DF = pd.read_csv( socialMetricsInputFolder+orgName+'_facebook.csv', index_col=0)
        tw_DF = pd.read_csv( socialMetricsInputFolder+orgName+'_twitter.csv', index_col=0)
        # Read in bitly data
        bitly_tw = pd.read_csv( bitlyInputFolder+'1. extractedBitly/'+orgName+'_twitter.csv', index_col=0)
        bitly_fb = pd.read_csv( bitlyInputFolder+'1. extractedBitly/'+orgName+'_facebook.csv', index_col=0)
        # Read in sentiment data
        sent_tw = pd.read_csv( sentimentInputFolder+'1. extractedSentiment/'+orgName+'_twitter.csv', index_col=0)
        sent_fb = pd.read_csv( sentimentInputFolder+'1. extractedSentiment/'+orgName+'_facebook.csv', index_col=0)
        # Read in merged articles' urls
        mergedUrls_DF = pd.read_csv( mergedUrlsInputFolder+orgName+'_merged.csv', index_col=0)
        mergedUrls_DF = mergedUrls_DF.drop_duplicates(cols='TW_matchUrl', take_last=False)
        
        #Merging fb and tw social data with merged articles' urls
        tw_DF_summed = sumSocialMetrics('twitter', tw_DF, mergedUrls_DF, destFolder, orgName)
        fb_DF_summed = sumSocialMetrics('facebook', fb_DF, mergedUrls_DF, destFolder, orgName)
        finalmerged_DF = tw_DF_summed.merge(mergedUrls_DF, left_on='tw_expandedUrl', right_on='TW_ShortenedUrl', how="inner")
        finalmerged_DF = finalmerged_DF.merge(fb_DF_summed, left_on='FB_ShortenedUrl', right_on='fb_expandedUrl', how="inner")
        
        # Merge in Bitly
        finalmerged_DF = finalmerged_DF.merge(bitly_tw, left_on='TW_ShortenedUrl', right_on='tw_url', how="left", left_index=True)
        finalmerged_DF = finalmerged_DF.merge(bitly_fb, left_on='FB_ShortenedUrl', right_on='fb_url', how="left", left_index=True)
        
        # Merge in Sentiment
        finalmerged_DF = finalmerged_DF.merge(sent_tw, left_on='tw_id', right_on='tw_id', how="left", copy=False)
        finalmerged_DF = finalmerged_DF.merge(sent_fb, left_on='fb_id', right_on='fb_id', how="left", copy=False)
        
        # Rename column names
        finalmerged_DF.rename(columns={'tw_article_appearance_Count_x':'tw_article_appearance_Count', 
                                       'FB_HistoryUrls_x':'FB_HistoryUrls',
                                       'fb_article_appearance_Count_x':'fb_article_appearance_Count',
                                       'tw_timeRetrieved_x':'tw_timeRetrieved',
                                       'fb_timeRetrieved_x':'fb_timeRetrieved',
                                       'tw_sentiment_positive_x':'tw_sentiment_positive'}, inplace=True)
        # To csv
        finalmerged_DF.to_csv(destFolder+orgName+'_finaltable.csv', encoding='utf-8')
        rowCount += finalmerged_DF.shape[0]
        print finalmerged_DF.shape
        tw_repostCounts += finalmerged_DF[finalmerged_DF['tw_article_appearance_Count'] > 1].count()['tw_expandedUrl']
        fb_repostCounts += finalmerged_DF[finalmerged_DF.fb_article_appearance_Count > 1].count()['tw_expandedUrl']
        
        
    print "Number of articles: %d" % rowCount
    print "Number of tw reposts: %d" % tw_repostCounts
    print "Number of fb reposts: %d" % fb_repostCounts
    return fb_DF, tw_DF, mergedUrls_DF, finalmerged_DF


# BBC and NBCnews Example
test_fb_DF, test_tw_DF, test_mergedUrls_DF, test_finalmerged_DF = mergeSocialMetricsData(pathMergedUrls, pathSocialMediaMetrics, 
                                                                                         pathBitly, pathSentiment, [organizationNames[3], organizationNames[12]], pathFinalTable)

#Full Run
fb_DF, tw_DF, mergedUrls_DF, finalmerged_DF = mergeSocialMetricsData(pathMergedUrls, pathSocialMediaMetrics, 
                                                                                         pathBitly, pathSentiment, organizationNames, pathFinalTable)
